requirements.txt
asgiref==3.7.2
cloudinary==1.33.0
crispy-bootstrap5==0.7
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.20
django-allauth==0.54.0
django-crispy-forms==1.14.0
django-summernote==0.8.20.0
gunicorn==21.2.0
oauthlib==3.2.2
psycopg2==2.9.6
PyJWT==2.8.0
python-slugify==8.0.1
python3-openid==3.2.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
urllib3==1.26.16
whitenoise==6.5.0
